+++ Cargo.toml
@@ -1,3 +1,3 @@
 [workspace]
-members = ["client", "signaling"]
+members = ["client", "signaling", "media"]
 resolver = "2"
