+++ signaling/Cargo.toml
@@ -3,16 +3,11 @@ name = "signaling"
 version = "0.1.0"
 edition = "2021"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
 [dependencies]
 anyhow = "1.0.95"
 axum = { version = "0.8.1", features = [] }
 axum-extra = { version = "0.10.1" }
 axum-server = { version = "0.7.2", features = ["tls-rustls"] }
-gstreamer = "0.23.5"
-gstreamer-app = "0.23.5"
-gstreamer-base = "0.23.6"
 once_cell = "1.20.2"
 reqwest = { version = "0.12.9", features = ["blocking", "json"] }
 rustls = "0.23.19"
