+++ signaling/src/lib.rs
@@ -1,8 +1,3 @@
 pub mod client;
 pub mod message;
 pub mod util;
-
-pub enum WebRtcEvent {
-    Continue,
-    Disconnected,
-}
